+++ fixer-app-backend/src/main/java/org/example/fixerappbackend/repo/UsuarioRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepo extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepo extends JpaRepository<Usuario, Long> {
     Usuario findByEmail(String email);
 
     Usuario findByNombreUsuario(String nombre);
